+++ Londono_Camilo_Morales_Miguel_S25 - Evidencia de Aprendizaje Unidad 2/CoopRKC/src/modelo/FondosInsuficientesException.java
@@ -1,0 +1,7 @@
+package modelo;
+
+public class FondosInsuficientesException extends Exception {
+    public FondosInsuficientesException(String message) {
+        super(message);
+    }
+}
