+++ Londono_Camilo_Morales_Miguel_S25 - Evidencia de Aprendizaje Unidad 2/CoopRKC/src/modelo/Transaccion.java
@@ -1,0 +1,6 @@
+package modelo;
+
+public interface Transaccion {
+    void ejecutar() throws Exception;
+    double getMonto();
+}
