--- Londono_Camilo_Morales_Miguel_S25 - Evidencia de Aprendizaje Unidad 2/Londono_Camilo_Morales_Miguel_S25 - Evidencia de Aprendizaje Unidad 2/miguel/src/modelo/MontoInvalidoException.java
@@ -1,7 +1,0 @@
-package modelo;
-
-public class MontoInvalidoException extends Exception {
-    public MontoInvalidoException(String message) {
-        super(message);
-    }
-}
