--- Londono_Camilo_Morales_Miguel_S25 - Evidencia de Aprendizaje Unidad 2/Londono_Camilo_Morales_Miguel_S25 - Evidencia de Aprendizaje Unidad 2/miguel/src/modelo/Transaccion.java
@@ -1,6 +1,0 @@
-package modelo;
-
-public interface Transaccion {
-    void ejecutar() throws Exception;
-    double getMonto();
-}
